requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
astroid==1.6.6
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/test.in
    #   pylint
attrs==24.2.0
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   referencing
certifi==2024.8.30
    # via
    #   -r requirements/base.txt
    #   elasticsearch
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
chardet==3.0.4
    # via
    #   -r requirements/base.txt
    #   diff-cover
    #   requests
click==8.1.7
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==1.8.0
    # via -r requirements/test.in
colorama==0.4.6
    # via pylint
coverage[toml]==7.6.1
    # via
    #   -r requirements/test.in
    #   pytest-cov
cryptography==43.0.1
    # via
    #   -r requirements/base.txt
    #   pyjwt
ddt==1.7.2
    # via -r requirements/test.in
diff-cover==9.2.0
    # via -r requirements/test.in
distlib==0.3.8
    # via virtualenv
    # via
    #   -c requirements/common_constraints.txt
    #   -r requirements/base.txt
    #   django-cors-headers
    #   django-crum
    #   django-nine
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   drf-spectacular
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-drf-extensions
django-cors-headers==3.14.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-elasticsearch-dsl==7.4
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   django-elasticsearch-dsl-drf
django-elasticsearch-dsl-drf==0.22.5
    # via -r requirements/base.txt
django-nine==0.2.7
    # via
    #   -r requirements/base.txt
    #   django-elasticsearch-dsl-drf
django-waffle==4.1.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.15.2
    # via
    #   -r requirements/base.txt
    #   django-elasticsearch-dsl-drf
    #   drf-jwt
    #   drf-spectacular
    #   edx-drf-extensions
dnspython==2.6.1
    # via
    #   -r requirements/base.txt
    #   pymongo
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
drf-spectacular==0.27.2
    # via -r requirements/base.txt
edx-django-release-util==1.4.0
    # via -r requirements/base.txt
edx-django-utils==5.16.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-drf-extensions==10.4.0
    # via -r requirements/base.txt
edx-opaque-keys==2.11.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
elasticsearch==7.13.4
    # via
    #   -c requirements/common_constraints.txt
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   django-elasticsearch-dsl-drf
    #   elasticsearch-dsl
elasticsearch-dsl==7.4.1
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   django-elasticsearch-dsl
    #   django-elasticsearch-dsl-drf
factory-boy==3.3.1
    # via -r requirements/test.in
faker==30.1.0
    # via factory-boy
filelock==3.16.1
    # via
    #   tox
    #   virtualenv
gunicorn==23.0.0
    # via -r requirements/base.txt
idna==2.10
    # via
    #   -r requirements/base.txt
    #   requests
inflection==0.5.1
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via
    #   code-annotations
    #   diff-cover
jsonschema==4.23.0
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
jsonschema-specifications==2023.12.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
lazy-object-proxy==1.10.0
    # via astroid
markupsafe==2.1.5
    # via jinja2
more-itertools==5.0.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/test.in
mysqlclient==2.2.4
    # via -r requirements/base.txt
newrelic==10.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
packaging==24.1
    # via
    #   -r requirements/base.txt
    #   django-nine
    #   gunicorn
    #   pytest
    #   tox
path-py==9.1
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
pbr==6.1.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pep8==1.7.1
    # via -r requirements/test.in
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via
    #   diff-cover
    #   pytest
    #   tox
psutil==6.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
py==1.11.0
    # via tox
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pygments==2.18.0
    # via diff-cover
pyjwt[crypto]==2.9.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
pylint==1.5.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/test.in
pymongo==4.10.1
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==8.3.3
    # via
    #   -r requirements/test.in
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements/test.in
pytest-django==4.9.0
    # via -r requirements/test.in
python-dateutil==2.4.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   elasticsearch-dsl
    #   faker
python-slugify==8.0.4
    # via code-annotations
pytz==2024.2
    # via -r requirements/base.txt
pyyaml==6.0.2
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   drf-spectacular
    #   edx-django-release-util
referencing==0.35.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   jsonschema-specifications
requests==2.25.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   edx-drf-extensions
rpds-py==0.20.0
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   referencing
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   astroid
    #   django-elasticsearch-dsl
    #   django-elasticsearch-dsl-drf
    #   edx-django-release-util
    #   elasticsearch-dsl
    #   more-itertools
    #   pylint
    #   python-dateutil
    #   tox
sqlparse==0.5.1
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.3.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tox==3.28.0
    # via -r requirements/test.in
typing-extensions==4.12.2
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
    #   faker
uritemplate==4.1.1
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
urllib3==1.26.20
    # via
    #   -r requirements/base.txt
    #   elasticsearch
    #   requests
virtualenv==20.26.6
    # via tox
wrapt==1.16.0
    # via astroid
